[geometry-merger] Add property for keeping colors when material color is used
When merged entities are using material colors for their whole mesh
(as opposed to vertex or faces colors), the resulting merged mesh
doesn't show those colors. This patch addes a new property,
materialColors, that when true, copies the material color
(if vertexColors is THREE.NoColors) to the faces in the mesh of
the entity to be merged, so that when merged, the corresponding
faces keep the color. This allows the resulting mesh to have the
coloring appearence of the original entities, and is quite useful
when AFrame materials with a single color are used.