Fix look-at component for vec3 targets and cameras
fix supermedium/superframe#229

Vec3 targets were not working. As weddingdj observed, in the update
method the `data` property was the default empty string, and not
the component attribute. Not sure the underlying cause, but probably
in A-Frame's component `buildData` method.

This results in the component immediately removing itself.

Resolved here by changing the default to a vec3 rather than empty
string.

Might also resolve supermedium/superframe#217

Also fixed the -z flip with cameras. The line of code that performs
the flip was commented out, and it is unclear why. I restored that
line and moved the logic to a `lookAt` method on the component, so
it could be used in the `update` path as well, which never properly
accounted for the camera.

In addition, hooked up events to watch for the camera being attached
or detached from the component, which at the very least makes this
component's camera behavior independent of the attribute order.

Finally, I edited the demo scene to add uses of the component that
execute these code paths, to verify it works.