WebXR compatibility update for index.js
This update makes thumb-controls fully compatible with WebXR. In May 2020 an update to the Oculus Browser rendered thumb-controls inoperable on Oculus Quest and Oculus Go. This update fixes that. The WebXR Gamepad module revealed an axis 2 and an axis 3 which we added via an else statement on lines 188 and 189. The WebXR Gamepad API also refers to "touchpad" so "touchpad event listeners were created. Also Axis 2 and 3 were added to the GetDistance and GetAngle functions respectively via an else statement.